Rebase against latest ocis driver

+++ pkg/storage/fs/s3ng/recycle.go
@@ -50,13 +50,14 @@ func (fs *s3ngfs) ListRecycle(ctx context.Context) (items []*provider.RecycleIte
 	items = make([]*provider.RecycleItem, 0)
 
 	// TODO how do we check if the storage allows listing the recycle for the current user? check owner of the root of the storage?
+	// use permissions ReadUserPermissions?
 	if fs.o.EnableHome {
 		if !ownerPermissions.ListContainer {
 			log.Debug().Msg("owner not allowed to list trash")
 			return items, errtypes.PermissionDenied("owner not allowed to list trash")
 		}
 	} else {
-		if !defaultPermissions.ListContainer {
+		if !noPermissions.ListContainer {
 			log.Debug().Msg("default permissions prevent listing trash")
 			return items, errtypes.PermissionDenied("default permissions prevent listing trash")
 		}
@@ -232,14 +233,15 @@ func (fs *s3ngfs) PurgeRecycleItem(ctx context.Context, key string) (err error) 
 }
 
 func (fs *s3ngfs) EmptyRecycle(ctx context.Context) error {
+	u, ok := user.ContextGetUser(ctx)
 	// TODO what permission should we check? we could check the root node of the user? or the owner permissions on his home root node?
-	// The current impl will wipe your own trash. or when enable home is false the trash of the 'root'
-	if fs.o.EnableHome {
-		u := user.ContextMustGetUser(ctx)
-		// TODO use layout, see Tree.Delete() for problem
-		return os.RemoveAll(filepath.Join(fs.o.Root, "trash", u.Id.OpaqueId))
+	// The current impl will wipe your own trash. or when no user provided the trash of 'root'
+	if !ok {
+		return os.RemoveAll(fs.getRecycleRoot(ctx))
 	}
-	return os.RemoveAll(fs.getRecycleRoot(ctx))
+
+	// TODO use layout, see Tree.Delete() for problem
+	return os.RemoveAll(filepath.Join(fs.o.Root, "trash", u.Id.OpaqueId))
 }
 
 func getResourceType(isDir bool) provider.ResourceType {
